feat: add 'Share to all workspaces' button & separate to helper (#4146)
* chore: fix minor style issue


* chore: update language


* refactor: separate code to helper


* feat: add 'Share to all workspaces' button


* chore: hide share menu in deprecated dashboards


---------
