+++ apps/web/src/common/modules/widgets/_components/WidgetFrame.vue
@@ -255,6 +255,7 @@ const handleToggleWidth = () => {
         .metadata-content {
             @apply text-label-md;
             display: flex;
+            line-height: unset;
             flex-direction: column;
             gap: 0.5rem;
             min-height: 6.25rem;
